0 - Projects/Flask Rest Assignment with JWT/requirements.txt
Flask = 2.0.3
pyJWT = 2.3.0
Werkzeug = 2.0.3